aircraft-tracking-system/requirements.txt
# Aircraft Tracking and Maintenance System
# Real-time monitoring and maintenance tracking solution

# Core
python-dateutil>=2.8.2
pytz>=2023.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Real-time Communication
websockets>=12.0
python-socketio>=5.10.0
aiohttp>=3.9.0

# Embedded/Hardware Integration
pyserial>=3.5
paho-mqtt>=1.6.1
RPi.GPIO>=0.7.1; platform_system=="Linux"

# Monitoring & Alerts
prometheus-client>=0.18.0
python-telegram-bot>=20.6
twilio>=8.10.0

# Scheduling
apscheduler>=3.10.4

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
structlog>=23.2.0
